seperates post sign up page into components
